+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_verifyattestationtimeouts.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "173666b461e7ec49380b08095ac1b1d1";
+module.exports.hash = "1eb4941133e557d3c27e29544fbe6510";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}