+++ sdk/attestation/attestation/src/attestationAdministrationClient.ts
@@ -155,10 +155,13 @@ export class AttestationAdministrationClient {
       const token = new AttestationTokenImpl(getPolicyResult.token);
 
       // Validate the token returned from the service.
-      token.validateToken(
+      const problems = token.getTokenProblems(
         await this.signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
+      if (problems.length) {
+        throw new Error(problems.join(";"));
+      }
 
       // Deserialize the PolicyResult object to retrieve the underlying policy
       //  token
@@ -194,8 +197,6 @@ export class AttestationAdministrationClient {
    *
    * @param attestationType - Attestation Type for which to set policy.
    * @param newPolicyDocument - Policy document to be set.
-   * @param privateKey - optional private key used to sign the policy document.
-   * @param certificate - optional certificate used to verify the policy document.
    * @param options - call options.
    * @returns An {@link AttestationResponse} wrapping a {@link PolicyResult}.
    *  Clients can use the PolicyResult to validate that the policy was actually
@@ -213,30 +214,33 @@ export class AttestationAdministrationClient {
   public async setPolicy(
     attestationType: AttestationType,
     newPolicyDocument: string,
-    privateKey?: string,
-    certificate?: string,
-    options: AttestationAdministrationClientOperationOptions = {}
+    options: AttestationAdministrationClientOperationOptions & {
+      privateKey?: string;
+      certificate?: string;
+    } = {}
   ): Promise<AttestationResponse<PolicyResult>> {
     const { span, updatedOptions } = createSpan(
       "AttestationAdministrationClient-setPolicy",
       options
     );
     try {
-      if ((!privateKey && certificate) || (privateKey && !certificate)) {
+      if (
+        (!options.privateKey && options.certificate) ||
+        (options.privateKey && !options.certificate)
+      ) {
         throw new Error(
           "If privateKey is specified, certificate must also be provided. If certificate is provided, privateKey must also be provided."
         );
       }
 
-      if (privateKey && certificate) {
-        verifyAttestationSigningKey(privateKey, certificate);
+      if (options.privateKey && options.certificate) {
+        verifyAttestationSigningKey(options.privateKey, options.certificate);
       }
 
       const storedAttestationPolicy = new StoredAttestationPolicy(newPolicyDocument).serialize();
       const setPolicyToken = AttestationTokenImpl.create({
         body: storedAttestationPolicy,
-        privateKey: privateKey,
-        certificate: certificate
+        ...options
       });
 
       const setPolicyResult = await this._client.policy.set(
@@ -248,10 +252,13 @@ export class AttestationAdministrationClient {
       // The attestation token returned from the service has a PolicyResult
       // object as the body.
       const token = new AttestationTokenImpl(setPolicyResult.token);
-      token.validateToken(
+      const problems = token.getTokenProblems(
         await this.signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
+      if (problems.length) {
+        throw new Error(problems.join(";"));
+      }
 
       // Deserialize the PolicyResult object to retrieve the underlying policy
       //  token
@@ -274,8 +281,6 @@ export class AttestationAdministrationClient {
    * the default value.
    *
    * @param attestationType - Attestation Type for which to set policy.
-   * @param privateKey - optional private key used to sign the policy document
-   * @param certificate - optional certificate used to verify the policy document.
    * @param options - call options.
    * @returns An {@link AttestationResponse} wrapping a {@link PolicyResult}.
    *  Clients can use the PolicyResult to validate that the policy was actually
@@ -293,28 +298,32 @@ export class AttestationAdministrationClient {
 
   public async resetPolicy(
     attestationType: AttestationType,
-    privateKey?: string,
-    certificate?: string,
-    options: AttestationAdministrationClientOperationOptions = {}
+    options: AttestationAdministrationClientOperationOptions & {
+      privateKey?: string;
+      certificate?: string;
+    } = {}
   ): Promise<AttestationResponse<PolicyResult>> {
     const { span, updatedOptions } = createSpan(
       "AttestationAdministrationClient-setPolicy",
       options
     );
     try {
-      if ((!privateKey && certificate) || (privateKey && !certificate)) {
+      if (
+        (!options.privateKey && options.certificate) ||
+        (options.privateKey && !options.certificate)
+      ) {
         throw new Error(
           "If privateKey is specified, certificate must also be provided. If certificate is provided, privateKey must also be provided."
         );
       }
 
-      if (privateKey && certificate) {
-        verifyAttestationSigningKey(privateKey, certificate);
+      if (options.privateKey && options.certificate) {
+        verifyAttestationSigningKey(options.privateKey, options.certificate);
       }
 
       const resetPolicyToken = AttestationTokenImpl.create({
-        privateKey: privateKey,
-        certificate: certificate
+        privateKey: options.privateKey,
+        certificate: options.certificate
       });
 
       const resetPolicyResult = await this._client.policy.reset(
@@ -326,10 +335,13 @@ export class AttestationAdministrationClient {
       // The attestation token returned from the service has a PolicyResult
       // object as the body.
       const token = new AttestationTokenImpl(resetPolicyResult.token);
-      token.validateToken(
+      const problems = token.getTokenProblems(
         await this.signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
+      if (problems.length) {
+        throw new Error(problems.join(";"));
+      }
 
       // Deserialize the PolicyResult object to retrieve the underlying policy
       //  token
@@ -367,10 +379,13 @@ export class AttestationAdministrationClient {
       // The attestation token returned from the service has a PolicyResult
       // object as the body.
       const token = new AttestationTokenImpl(getCertificatesResult.token);
-      token.validateToken(
+      const problems = token.getTokenProblems(
         await this.signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
+      if (problems.length) {
+        throw new Error(problems.join(";"));
+      }
 
       // Deserialize the PolicyResult object to retrieve the underlying policy
       //  token
@@ -470,10 +485,13 @@ export class AttestationAdministrationClient {
       // The attestation token returned from the service has a PolicyResult
       // object as the body.
       const token = new AttestationTokenImpl(addCertificateResult.token);
-      token.validateToken(
+      const problems = token.getTokenProblems(
         await this.signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
+      if (problems.length) {
+        throw new Error(problems.join(";"));
+      }
 
       // Deserialize the PolicyCertificatesModificationResult object.
       const result = TypeDeserializer.deserialize(
@@ -585,10 +603,13 @@ export class AttestationAdministrationClient {
       // The attestation token returned from the service has a PolicyResult
       // object as the body.
       const token = new AttestationTokenImpl(removeCertificateResult.token);
-      token.validateToken(
+      const problems = token.getTokenProblems(
         await this.signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
+      if (problems.length) {
+        throw new Error(problems.join(";"));
+      }
 
       // Deserialize the PolicyCertificatesModificationResult object.
       const result = TypeDeserializer.deserialize(