+++ sdk/attestation/attestation/src/index.ts
@@ -6,7 +6,6 @@ export {
   AttestationClientOptions,
   AttestOpenEnclaveOptions,
   AttestSgxEnclaveOptions,
-  AttestSgxEnclaveBrowserOptions,
   AttestTpmOptions,
   AttestationClientOperationOptions
 } from "./attestationClient";